pyproject.toml
[project]
name = "sports-papertrader-web"
version = "0.1.0"
description = "Web dashboard for Polymarket paper trading system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
papertrader = "app.main:run"
